Added all the scenarios for CDNC-4316
