+++ test/replaytests/signatureBugWorkflow.go
@@ -1,0 +1,195 @@
+package replaytests
+
+import (
+	"fmt"
+	"go.uber.org/cadence/workflow"
+	"go.uber.org/zap"
+	"time"
+)
+
+// greetingWorkflow demonstrates the normal case where we are executing multiple acitivities and expect a certain result.
+// Replayer should Pass this.
+func greetingsWorkflow(ctx workflow.Context) error {
+	// Get Greeting.
+	ao := workflow.ActivityOptions{
+		ScheduleToStartTimeout: time.Minute,
+		StartToCloseTimeout:    time.Minute,
+		HeartbeatTimeout:       time.Second * 20,
+	}
+	ctx = workflow.WithActivityOptions(ctx, ao)
+
+	logger := workflow.GetLogger(ctx)
+	var greetResult string
+	err := workflow.ExecuteActivity(ctx, getGreetingActivity).Get(ctx, &greetResult)
+	if err != nil {
+		logger.Error("Get greeting failed.", zap.Error(err))
+		return err
+	}
+
+	// Get Name.
+	var nameResult string
+	err = workflow.ExecuteActivity(ctx, getNameActivity).Get(ctx, &nameResult)
+	if err != nil {
+		logger.Error("Get name failed.", zap.Error(err))
+		return err
+	}
+
+	// Say Greeting.
+	var sayResult string
+	err = workflow.ExecuteActivity(ctx, sayGreetingActivity, greetResult, nameResult).Get(ctx, &sayResult)
+	if err != nil {
+		logger.Error("Marshalling failed with error.", zap.Error(err))
+		return err
+	}
+
+	logger.Info("Workflow completed.", zap.String("Result", sayResult))
+	return nil
+}
+
+// greetingWorkflow demonstrates the case where an activity returns a response that contains the desired string but also returns unwanted objects.
+// Replayer is not able to catch this because the desired string is still being returned as recorded in History.
+func greetingsWorkflow2(ctx workflow.Context) error {
+	// Get Greeting.
+	ao := workflow.ActivityOptions{
+		ScheduleToStartTimeout: time.Minute,
+		StartToCloseTimeout:    time.Minute,
+		HeartbeatTimeout:       time.Second * 20,
+	}
+	ctx = workflow.WithActivityOptions(ctx, ao)
+
+	logger := workflow.GetLogger(ctx)
+	var greetResult string
+	err := workflow.ExecuteActivity(ctx, getGreetingActivity).Get(ctx, &greetResult)
+	if err != nil {
+		logger.Error("Get greeting failed.", zap.Error(err))
+		return err
+	}
+
+	// Get Name.
+	var nameResult string
+	err = workflow.ExecuteActivity(ctx, getNameActivity2).Get(ctx, &nameResult)
+	if err != nil {
+		logger.Error("Get name failed.", zap.Error(err))
+		return err
+	}
+
+	// Say Greeting.
+	var sayResult string
+	err = workflow.ExecuteActivity(ctx, sayGreetingActivity, greetResult, nameResult).Get(ctx, &sayResult)
+	if err != nil {
+		logger.Error("Marshalling failed with error.", zap.Error(err))
+		return err
+	}
+
+	logger.Info("Workflow completed.", zap.String("Result", sayResult))
+	return nil
+}
+
+// greetingWorkflow demonstrates the case where an activity returns a response that contains the desired string but also returns unwanted objects.
+// Replayer is not able to catch this because the desired string is still being returned as recorded in History.
+func greetingsWorkflow3(ctx workflow.Context) error {
+	// Get Greeting.
+	ao := workflow.ActivityOptions{
+		ScheduleToStartTimeout: time.Minute,
+		StartToCloseTimeout:    time.Minute,
+		HeartbeatTimeout:       time.Second * 20,
+	}
+	ctx = workflow.WithActivityOptions(ctx, ao)
+
+	logger := workflow.GetLogger(ctx)
+	var greetResult string
+	err := workflow.ExecuteActivity(ctx, getGreetingActivity).Get(ctx, &greetResult)
+	if err != nil {
+		logger.Error("Get greeting failed.", zap.Error(err))
+		return err
+	}
+
+	// Get Name.
+	var nameResult string
+	err = workflow.ExecuteActivity(ctx, getNameActivity3).Get(ctx, &nameResult)
+	if err != nil {
+		logger.Error("Get name failed.", zap.Error(err))
+		return err
+	}
+
+	// Say Greeting.
+	var sayResult string
+	err = workflow.ExecuteActivity(ctx, sayGreetingActivity, greetResult, nameResult).Get(ctx, &sayResult)
+	if err != nil {
+		logger.Error("Marshalling failed with error.", zap.Error(err))
+		return err
+	}
+
+	logger.Info("Workflow completed.", zap.String("Result", sayResult))
+	return nil
+}
+
+// greetingWorkflow demonstrates the case where an activity returns a response that contains the desired string but also returns unwanted objects.
+// Replayer is not able to catch this because the desired string is still being returned as recorded in History.
+func greetingsWorkflow4(ctx workflow.Context) error {
+	// Get Greeting.
+	ao := workflow.ActivityOptions{
+		ScheduleToStartTimeout: time.Minute,
+		StartToCloseTimeout:    time.Minute,
+		HeartbeatTimeout:       time.Second * 20,
+	}
+	ctx = workflow.WithActivityOptions(ctx, ao)
+
+	logger := workflow.GetLogger(ctx)
+	var greetResult int64
+	err := workflow.ExecuteActivity(ctx, getGreetingActivity).Get(ctx, &greetResult)
+	if err != nil {
+		logger.Error("Get greeting failed.", zap.Error(err))
+		return err
+	}
+
+	// Get Name.
+	var nameResult string
+	err = workflow.ExecuteActivity(ctx, getNameActivity4).Get(ctx, &nameResult)
+	if err != nil {
+		logger.Error("Get name failed.", zap.Error(err))
+		return err
+	}
+
+	// Say Greeting.
+	var sayResult string
+	err = workflow.ExecuteActivity(ctx, sayGreetingActivity, greetResult, nameResult).Get(ctx, &sayResult)
+	if err != nil {
+		logger.Error("Marshalling failed with error.", zap.Error(err))
+		return err
+	}
+
+	logger.Info("Workflow completed.", zap.String("Result", sayResult))
+	return nil
+}
+
+// Get Name Activity.
+func getNameActivity2() (string, int64, error) {
+	return "Cadence", 1, nil
+}
+
+// Get Name Activity.
+func getNameActivity3() (int64, error) {
+	return 1, nil
+}
+
+// Get Name Activity.
+func getNameActivity4() (int64, error) {
+	return 1, nil
+}
+
+// Get Name Activity.
+func getNameActivity() (string, error) {
+	return "Cadence", nil
+}
+
+// Get Greeting Activity.
+func getGreetingActivity() (string, error) {
+	return "Hello", nil
+}
+
+// Say Greeting Activity.
+func sayGreetingActivity(greeting string, name string) (string, error) {
+	result := fmt.Sprintf("Greeting: %s %s!\n", greeting, name)
+	return result, nil
+}
